An initial, untested, updated os_ironic_node module
Updated os_ironic_node to include support support for
basic power and maintenance state assertions, as well
as support for changes between Ansible 1.9.x and
current Ansible 2.0 development.

NOTE: This requires a number of updates to the
shade library that have not landed, and cannot be
used at this time.